+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## 1.16.2.dev (development stage/unreleased/unstable)
+### Added
+- restart if "with connection" in socket gets closed
 
 ## 1.16.2
 ### Fixed